+++ src/PalTracker/InMemoryTimeEntryRepository.cs
@@ -3,7 +3,7 @@ using System.Linq;
 namespace PalTracker{
     public class InMemoryTimeEntryRepository : ITimeEntryRepository
     {
-        private readonly IDictionary<int, TimeEntry> _inMemoryTimeEntry = new Dictionary<int, TimeEntry>(); 
+        private readonly IDictionary<long, TimeEntry> _inMemoryTimeEntry = new Dictionary<long, TimeEntry>(); 
 
         public TimeEntry Create(TimeEntry timeEntry){
             var inMemoryTimeEntryCountAsId = _inMemoryTimeEntry.Count+1;    
@@ -11,21 +11,21 @@ namespace PalTracker{
             _inMemoryTimeEntry.Add(inMemoryTimeEntryCountAsId, timeEntry);
             return timeEntry;
          }
-        public TimeEntry Find(int id){
+        public TimeEntry Find(long id){
             return _inMemoryTimeEntry[id];
         }
-        public bool Contains(int id){
+        public bool Contains(long id){
             return _inMemoryTimeEntry.ContainsKey(id);
         }
         public IEnumerable<TimeEntry> List(){
             return _inMemoryTimeEntry.Values.ToList();
         }
-        public TimeEntry Update(int id, TimeEntry timeEntry){
+        public TimeEntry Update(long id, TimeEntry timeEntry){
             timeEntry.Id = id;
             _inMemoryTimeEntry[id] = timeEntry;
             return timeEntry;
         }
-        public void Delete(int id){
+        public void Delete(long id){
             _inMemoryTimeEntry.Remove(id);
         }
     }
