+++ src/PalTracker/TimEntry.cs
@@ -5,13 +5,13 @@ namespace PalTracker
 
     public struct TimeEntry
     {
-        public int? Id { get; set;}
-        public int ProjectId { get; set; }
-        public int UserId { get; set; }
+        public long? Id { get; set;}
+        public long ProjectId { get; set; }
+        public long UserId { get; set; }
         public DateTime Date { get; set; }
         public int Hours { get; set; }
 
-        public TimeEntry(int id, int projectId, int userId, DateTime date, int hours)
+        public TimeEntry(long id, long projectId, long userId, DateTime date, int hours)
         {
             Id = id;
             ProjectId = projectId;
@@ -20,7 +20,7 @@ namespace PalTracker
             Hours = hours;
         }
 
-        public TimeEntry(int projectId, int userId, DateTime date, int hours)
+        public TimeEntry(long projectId, long userId, DateTime date, int hours)
         {
             Id = null;
             ProjectId = projectId;
