+++ src/PalTracker/TimeEntryController.cs
@@ -21,7 +21,7 @@ namespace PalTracker
         }
 
         [HttpGet("{id}", Name = "GetTimeEntry")]
-        public IActionResult Read(int id)
+        public IActionResult Read(long id)
         {
             return _repository.Contains(id) ? (IActionResult) Ok(_repository.Find(id)) : NotFound();
         }
@@ -33,13 +33,13 @@ namespace PalTracker
         }
 
         [HttpPut("{id}")]
-        public IActionResult Update(int id, [FromBody] TimeEntry timeEntry)
+        public IActionResult Update(long id, [FromBody] TimeEntry timeEntry)
         {
             return _repository.Contains(id) ? (IActionResult) Ok(_repository.Update(id, timeEntry)) : NotFound();
         }
 
         [HttpDelete("{id}")]
-        public IActionResult Delete(int id)
+        public IActionResult Delete(long id)
         {
             if (!_repository.Contains(id))
             {
